refactor: organize table component code
